Move folded reindex test to test_reindex
